Refactor DevSession into a class
